refactor: Integrate embedding service and update related references for improved functionality
